Use int
Change from long to int in the test for day 16.

+++ tests/AdventOfCode.Tests/Puzzles/Y2021/Day16Tests.cs
@@ -18,10 +18,10 @@ public sealed class Day16Tests : PuzzleTest
     [InlineData("620080001611562C8802118E34", 12)]
     [InlineData("C0015000016115A2E0802F182340", 23)]
     [InlineData("A0016C880162017C3686B18A3D4780", 31)]
-    public void Y2021_Day16_Decode_Returns_Correct_Value(string transmission, long expected)
+    public void Y2021_Day16_Decode_Returns_Correct_Value(string transmission, int expected)
     {
         // Act
-        long actual = Day16.Decode(transmission);
+        int actual = Day16.Decode(transmission);
 
         // Assert
         actual.ShouldBe(expected);
